All routing logic was moved into router.go
